fix generate
