+++ .task/checksum/generate-catalog-smart
@@ -1,1 +1,1 @@
-a7a48043595250b7e9af892e62149ee0
+1e15ee0a94a81c9b472cc0ce6fa3410
