+++ .task/checksum/generate-graphql-smart
@@ -1,1 +1,1 @@
-7c0be6b0b81ac84a157cf36048638d2e
+c024787dd22674f82940a3f58b59dcd2
