+++ .task/checksum/generate-openapi-smart
@@ -1,1 +1,1 @@
-ea6b375ad70ff3d16b671f9c17e8f609
+8efbbdb87ff2cd1a3256194f0c90c0
