+++ internal/graphapi/checksum/.schema_checksum
@@ -1,1 +1,1 @@
-112683320590ab8caa1c5348581af6302ebf0e349fff7f7daa1d4c33596d16e0
+03c2b3ff9ecc888b4308d39255694a0752ad33c6c51152ea87fcf3112db89464
