+++ internal/graphapi/testclient/checksum/.client_checksum
@@ -1,1 +1,1 @@
-6a79335709d19ceed28622c6dd6e88068ccbf4351a21e5318758934d4efe2c4d
+13c59ae5c9ebebe65a0e8d28b24d51eb83f2f2b6c549a7e70e8a2e872d43b4b3
